+++ app/services/answer_factory.rb
@@ -9,6 +9,7 @@ class AnswerFactory
     case question.contentful_type
     when "radios" then RadioAnswer.new
     when "short_text" then ShortTextAnswer.new
+    when "long_text" then LongTextAnswer.new
     end
   end
 end